app/requirements.txt
# FastAPI and Server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1

# Redis
redis==5.0.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2

# OpenAI
openai==1.10.0

# Configuration & Environment
python-dotenv==1.0.1
pydantic==2.5.3
pydantic-settings==2.1.0

# Utilities
httpx==0.26.0

# PDF generation
weasyprint==62.3
pydyf==0.11.0

# Task queue
celery[redis]==5.3.6

# Google Drive API
google-api-python-client==2.118.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
